// src/types/gestao-pessoas/prestador.ts
export interface PrestadorPJ {
  id: string;
  razaoSocial: string;
  nomeCompleto: string;
  cnpj: string;
  cpf: string;
  inscricaoEstadual?: string;
  inscricaoMunicipal?: string;
  endereco?: string;
  cep?: string;
  cidade?: string;
  estado?: string;
  telefone?: string;
  email?: string;
  descricaoAtividade?: string;
  valorPrestacaoServico?: number;
  ajudaCusto?: number;
  auxilioConvenioMedico?: number;
  ajudaAluguel?: number;
  valeRefeicao?: number;
  cafeManha?: number;
  cafeTarde?: number;
  almoco?: number;
  grauDeRisco?: number;
  tempoContrato?: string;
  veiculo?: boolean;
  celular?: boolean;
  alojamento?: boolean;
  folgaCampo?: string;
  periodoFerias?: string;
  quantidadeDiasFerias?: number;
  banco?: string;
  agencia?: string;
  conta?: string;
  tipoConta?: string;
  chavePix?: string;
  dataInicioContrato?: string;
  created_at?: string;
  updated_at?: string;
}